add sweet alert no perfil

+++ includes/perfil/excluir_conta.php
@@ -9,10 +9,10 @@ if (!isset($_SESSION['id'])) {
 
 $user_id = $_SESSION['id'];
 $tipo_usuario = $_SESSION['tipo_usuario'];
-
 $tabela = ($tipo_usuario === 'tutor') ? 'tutores' : 'cuidadores';
 
-if (isset($_POST['confirmar_exclusao'])) {
+if ($_SERVER['REQUEST_METHOD'] === 'POST' && isset($_POST['confirmar_exclusao'])) {
+
     $sql = "DELETE FROM $tabela WHERE id = ?";
     
     if ($stmt = $mysqli->prepare($sql)) {
@@ -21,7 +21,8 @@ if (isset($_POST['confirmar_exclusao'])) {
         
         if ($stmt->affected_rows > 0) {
             session_destroy();
-            header('Location: conta_excluida.php');
+            header('Location: /views/shared/conta_excluida.php');
+            exit();
         } else {
             echo "Erro ao tentar excluir a conta. Tente novamente mais tarde.";
         }
@@ -30,6 +31,8 @@ if (isset($_POST['confirmar_exclusao'])) {
     } else {
         echo "Erro de conexão com o banco de dados.";
     }
+} else {
+    echo "Ação inválida.";
 }
 
 $mysqli->close();
